+++ react/components/FilterBar/FilterTag.js
@@ -157,6 +157,7 @@ class FilterTag extends PureComponent {
       submitFilterLabel,
       newFilterLabel,
       isMobile,
+      disabled,
       testIds,
     } = this.props
     const { isMenuOpen, virtualStatement } = this.state
@@ -171,7 +172,6 @@ class FilterTag extends PureComponent {
     const shouldOmitVerb = isMoreOptions
       ? false
       : options[subject].verbs.length === 1
-
     // this is temporary just to assure backward compatibility
     const compatibleOptions = {}
     Object.keys(options).forEach(opt => {
@@ -213,19 +213,21 @@ class FilterTag extends PureComponent {
 
     return (
       <div
+        className={classNames('br2 ba b--solid pv1 dib c-on-base', {
+          pr4: !isEmpty && !isMoreOptions,
+          'hover-bg-muted-5 b--muted-4':
+            (alwaysVisible && isEmpty) || isMoreOptions,
+          'bg-action-secondary hover-bg-action-secondary b--action-secondary': !(
+            (alwaysVisible && isEmpty) ||
+            isMoreOptions
+          ),
+          'bg-transparent': !disabled && alwaysVisible && isEmpty,
+          'bg-disabled': disabled,
+        })}
         ref={this.filterMenuContainer}
         style={{
           ...(isMenuOpen && OPEN_MENU_STYLE),
         }}
-        className={`br2 ba b--solid ${
-          isEmpty || isMoreOptions ? '' : 'pr4'
-        } pv1 dib ${
-          alwaysVisible && isEmpty
-            ? 'bg-transparent hover-bg-muted-5 b--muted-4'
-            : isMoreOptions
-            ? 'hover-bg-muted-5 b--muted-4'
-            : 'bg-action-secondary hover-bg-action-secondary b--action-secondary'
-        } c-on-base`}
       >
         <div className="flex items-stretch">
           <Menu
@@ -237,8 +239,15 @@ class FilterTag extends PureComponent {
               <button
                 data-testid={options[subject] && options[subject].testId}
                 type="button"
-                className="bw1 ba br2 v-mid relative bg-transparent b--transparent c-action-primary pointer w-100 outline-0"
+                className={classNames(
+                  'bw1 ba br2 v-mid relative b--transparent w-100 outline-0',
+                  {
+                    'bg-transparent c-action-primary pointer': !disabled,
+                    'bg-disabled': disabled,
+                  }
+                )}
                 onClick={isMenuOpen ? this.handleCloseMenu : this.openMenu}
+                disabled={disabled}
               >
                 <div className="flex items-center justify-center h-100 ph3 ">
                   <span className="flex items-center nl1 nowrap">
@@ -348,6 +357,7 @@ class FilterTag extends PureComponent {
 FilterTag.defaultProps = {
   alwaysVisible: false,
   isMoreOptions: false,
+  disabled: false,
   subjectPlaceholder: '…',
   newFilterLabel: 'New filter',
 }
@@ -366,6 +376,7 @@ FilterTag.propTypes = {
   newFilterLabel: PropTypes.string,
   device: PropTypes.string,
   isMobile: PropTypes.bool,
+  disabled: PropTypes.bool,
   testIds: PropTypes.shape({
     moreOptionsButton: PropTypes.string,
     submitFiltersButton: PropTypes.string,